attestations/finalProject
- Структура проекта создана в соответствии с паттернами MVC и Repository;
- Применена технология JPA (Hibernate) на слое работы с данными.
- В проекте использована дополнительная библиотека Lombok.
- Слой сервисов содержит полный набор CRUD операции для работы с таблицами базы данных проекта Stylefit.
- Дополнительный функционал вынесен в папку utils.
- В веб-контроллере приложения реализованы страницы для работы с приложением.
- Приложение принимает файлы в формате JSON.
- Удаление данных реализовано через Soft Delete флаг.
- Создан слой DTO для работы с приложением.
- Добавлена автоматическая генерация миграций БД;
- Проект покрыт unit-тестами
- В Maven добавлен Swagger UI.
Страницы веб-контроллера расписаны с помощью Swagger UI.
- Добавлена страница авторизации для входа в
приложение.
- Работа приложения проверена через Postman и Insomnia
