+++ src/main/java/ru/innopolis/StyleFit/models/Client.java
@@ -57,3 +57,6 @@ public class Client {
         return firstName + " " + lastName;
     }
 }
+
+
+
