+++ src/main/java/ru/innopolis/StyleFit/service/AuthService.java
@@ -13,10 +13,13 @@ public class AuthService {
 
     static {
         // Добавляем тестового пользователя (в реальном приложении используйте хеширование паролей)
-        users.put("testUser ", "password123"); // username: testUser , password: password123
+        users.put("testUser ", "password123"); // username: testUser, password: password123
     }
 
     public boolean authenticate(String username, String password) {
         return users.containsKey(username) && users.get(username).equals(password);
     }
 }
+
+
+
