+++ src/main/java/ru/innopolis/StyleFit/service/ResultAnalysisService.java
@@ -26,3 +26,7 @@ public class ResultAnalysisService {
 
     // Дополнительные методы для анализа результатов можно добавить здесь
 }
+
+
+
+
