+++ src/test/java/ru/innopolis/Stylefit/StylefitApplicationTests.java
@@ -82,152 +82,3 @@ class StylefitApplicationTests {
         }
     }
 }
-
-@SpringBootTest(classes = Application.class)
-
-class AuthServiceTest {
-
-    @Autowired
-    private AuthService authService;
-
-    @Test
-    void testAuthenticate_ValidUser () {
-        boolean isAuthenticated = authService.authenticate("testUser ", "password123");
-        assertThat(isAuthenticated).isTrue();
-    }
-
-    @Test
-    void testAuthenticate_InvalidUser () {
-        boolean isAuthenticated = authService.authenticate("testUser ", "wrongPassword");
-        assertThat(isAuthenticated).isFalse();
-    }
-
-    @Test
-    void testAuthenticate_NonExistingUser () {
-        boolean isAuthenticated = authService.authenticate("nonExistingUser ", "password123");
-        assertThat(isAuthenticated).isFalse();
-    }
-}
-
-
-@SpringBootTest(classes = Application.class) // Use this if you need the Spring context
-class ClientServiceTest {
-
-    @MockBean // Use @MockBean to mock the ClientService if using Spring context
-    private ClientService clientService;
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    public void testGetClientById() {
-        // Your test logic here
-    }
-
-    @Test
-    public void testGetAllClients() {
-        // Your test logic here
-    }
-
-    @Test
-    public void testDeleteClient() {
-        // Your test logic here
-    }
-}
-
-class NotificationServiceTest {
-
-    @InjectMocks
-    private NotificationService notificationService;
-
-    @Mock
-    private Athlete athlete;
-
-    @Mock
-    private Result result;
-
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
-    }
-
-    @Test
-    public void testSendResultsUpdate_WithValidResult() {
-        // Настройка
-        when(athlete.getFirstName()).thenReturn("Иван");
-        when(result.getTotalWorkouts()).thenReturn(5.0); // Возвращаем Double
-
-        // Вызов метода
-        notificationService.sendResultsUpdate(athlete, result);
-
-        // Проверка
-        // В данном случае можно использовать System.out или другой механизм для проверки вывода
-    }
-
-
-    @Test
-    public void testSendResultsUpdate_WithNullResult() {
-        // Настройка
-        when(athlete.getFirstName()).thenReturn("Иван");
-
-        // Вызов метода
-        notificationService.sendResultsUpdate(athlete, null);
-
-        // Проверка
-        // Здесь также можно проверить вывод или поведение
-    }
-
-    // Здесь можно добавить дополнительные тесты для метода sendWorkoutReminder
-}
-
-@SpringBootTest(classes = Application.class) // Укажите ваш главный класс приложения
-class ResultAnalysisServiceTest {
-
-    @Autowired
-    private ResultAnalysisService resultAnalysisService;
-
-    @MockBean
-    private ResultRepository resultRepository;
-    @Test
-    public void testCalculateAverageResults_WithResults() {
-        // Настройка мока
-        Long athleteId = 1L;
-        Athlete athlete = new Athlete(); // Создайте объект Athlete, если нужно, или используйте мока
-        athlete.setId(athleteId); // Убедитесь, что у спортсмена установлен идентификатор
-
-        List<Result> results = Arrays.asList(
-                new Result(1L, athlete, "Event1", 5.0, new Date()), // Убедитесь, что передаете правильные параметры
-                new Result(2L, athlete, "Event2", 10.0, new Date()),
-                new Result(3L, athlete, "Event3", 15.0, new Date())
-        );
-
-        // Настройка мока
-        when(resultRepository.findByAthlete_Id(athleteId)).thenReturn(results);
-
-        // Вызов метода
-        double average = resultAnalysisService.calculateAverageResults(athleteId);
-
-        // Отладочная информация
-        System.out.println("Average calculated: " + average); // Отладочное сообщение
-
-        // Проверка
-        assertEquals(10.0, average, 0.01); // Проверяем, что среднее значение равно 10.0
-    }
-
-
-
-    @Test
-    public void testCalculateAverageResults_WithoutResults() {
-        // Настройка мока
-        Long athleteId = 1L;
-        when(resultRepository.findByAthlete_Id(athleteId)).thenReturn(Collections.emptyList());
-
-        // Вызов метода
-        double average = resultAnalysisService.calculateAverageResults(athleteId);
-
-        // Проверка
-        assertEquals(0.0, average, 0.01); // Проверяем, что среднее значение равно 0.0
-    }
-}
-
